refactor: Member 서비스 인터페이스 제거(#10)
- Auth 서비스 인터페이스를 서비스 구현 클래스로 통합
- Member 서비스 인터페이스를 서비스 구현 클래스로 통합
