+++ src/main/java/com/zero/bwtableback/member/service/MemberService.java
@@ -1,4 +1,14 @@
 package com.zero.bwtableback.member.service;
 
-public interface MemberService {
+import com.zero.bwtableback.member.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class MemberService {
+
+    private final MemberRepository memberRepository;
+
+
 }
