--- src/main/java/com/zero/bwtableback/member/service/MemberServiceImpl.java
@@ -1,16 +1,0 @@
-package com.zero.bwtableback.member.service;
-
-import com.zero.bwtableback.member.repository.MemberRepository;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Service;
-
-@Service
-public class MemberServiceImpl implements MemberService {
-
-    private final MemberRepository memberRepository;
-
-    public MemberServiceImpl(MemberRepository memberRepository, BCryptPasswordEncoder passwordEncoder) {
-        this.memberRepository = memberRepository;
-    }
-
-}
